#86 standardize folder and file names
