+++ src/app-components/location-detail/header/location-detail-header.js
