+++ src/app-components/location-detail/location-detail.css
