+++ src/app-containers/SearchBox/SearchBox.test.js
@@ -4,7 +4,7 @@ import {render, screen, fireEvent} from '@testing-library/react'
 import userEvent from '@testing-library/user-event'
 
 // Component import:
-import SearchBox from "./index";
+import SearchBox from "./searchbox";
 
 // Utils:
 import { findByTestAttr } from "../../testUtils";
