+++ src/app-pages/Map/containers/map-details/map-details.css
