+++ src/app-pages/Map/map-page.js
@@ -1,0 +1,24 @@
+import React from 'react';
+import MapNavBar from './containers/map-navbar/map-navbar';
+import MapDetails from './containers/map-details/map-details';
+import Map from '../../app-containers/map/map';
+
+const MapPage = (props) => {
+  const opts = { center: [-95, 38.895], zoom: 5 };
+
+	return (
+		<main>
+			<MapNavBar />
+			<div className=" map-and-details-container " style={{display:"flex",flexDirection:"row"}}>
+				<div className="map-details" style={{padding:"0", flexGrow: 1}}>
+					<MapDetails/>
+				</div>
+				<div className="map-container" style={{padding:"0", flexGrow: 35}} >
+					<Map height={'900px'} options={opts} mapKey={"maps"}/>
+				</div>
+			</div>
+		</main>
+	);
+};
+
+export default MapPage;
