+++ src/app-pages/Reports/index.js
@@ -1,5 +1,5 @@
 import React from "react";
-import ReportsContainer from "../../app-components/Reports";
+import ReportsContainer from "../../app-components/reports/reports-container";
 
 const ReportsPage = () => {
